refactor: modularize extrinsic and simplify the data structure, removed unnecessary types
